requirements.txt
asgiref==3.8.1
atpublic==5.1
attrs==25.3.0
Django==4.0
mysqlclient==2.2.7
pillow==11.1.0
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
utils==1.0.2
